src/charity_partnerships.py
```python
from src.utils import CharitySchema
from src.config import charityData

def partnerWithCharities():
    """
    Function to handle charity partnerships and fundraising.
    """
    # Fetch the charity data
    charities = charityData

    # Loop through each charity
    for charity in charities:
        # Validate the charity data
        validated_charity = CharitySchema().load(charity)

        # If the charity data is valid
        if validated_charity:
            # Start a fundraising campaign
            startFundraisingCampaign(validated_charity)

def startFundraisingCampaign(charity):
    """
    Function to start a fundraising campaign for a charity.
    """
    # Announce the fundraising campaign
    print(f"Starting a fundraising campaign for {charity['name']}!")

    # Start the fundraising campaign
    # This is a placeholder for the actual implementation
    pass

def updateCharityData(new_charity_data):
    """
    Function to update the charity data.
    """
    # Validate the new charity data
    validated_charity = CharitySchema().load(new_charity_data)

    # If the new charity data is valid
    if validated_charity:
        # Update the charity data
        charityData.append(validated_charity)

        # Announce the update
        print(f"Updated charity data with a new charity: {validated_charity['name']}!")
```